// packages/shared/src/enums.ts
export enum PurchaseState {
  RECEIVED = 'RECEIVED',
  ORDER_REQUESTED = 'ORDER_REQUESTED',
  RESERVE_DELIVERED = 'RESERVE_DELIVERED',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
}

export enum OrderState {
  PENDING = 'PENDING',
  PROCESSING = 'PROCESSING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
}

export enum DeliveryStatus {
  PENDING = 'PENDING',
  BOOKED = 'BOOKED',
  IN_TRANSIT = 'IN_TRANSIT',
  DELIVERED = 'DELIVERED',
  FAILED = 'FAILED',
}